Analyse the PHP, don't just syntax-check it
Every other language here had a real analyser — dart analyze --fatal-infos,
flake8/pylint/mypy, shellcheck — while the app subtree had only `php -l`,
which would not notice a null dereference or a wrong argument type. Psalm at
level 2 and nextcloud/coding-standard now run in `dev lint` and CI.

No baseline, deliberately: a baseline records the mistakes that existed when
it was written and permits them for ever. The three suppressions are scoped
and carry their reasons — #[\Override] needs PHP 8.3 and info.xml declares
8.2; MissingDependency is confined to the one file holding an IRootFolder,
whose OCP interface extends a class the stubs do not ship; and
PropertyNotSetInConstructor is confined to tests/, where setUp() is the
de-facto constructor.

Stubs are pinned to Nextcloud 32, the lowest version info.xml claims, so an
API added in 33 or 34 is a finding here rather than a broken install.

Of 59 first-run issues, 13 were substantive: dailyHeartRate built per-day
lists and reduced them three times, ManifestReader tested one expression and
returned another, weight() indexed $series[-1] on an empty series, and
folderChildren() promised a list while returning a keyed array. The rest
were implicit int/float coercions, now visible, plus ten classes nothing
extends made final.

+++ nextcloud_app/tests/Unit/Sleep/NightReconcilerTest.php
@@ -29,7 +29,7 @@ final class NightReconcilerTest extends TestCase {
 		return new SleepEpisodeReading(
 			start: Readings::at($start),
 			end: Readings::at($end),
-			totalSleepMillis: (int)round($totalMinutes * self::MINUTE),
+			totalSleepMillis: (int)round($totalMinutes * (float)self::MINUTE),
 			isMainSleep: true,
 			awakenings: 99,
 		);
